fix add duplicate category

+++ controller/handleAddCategory.php
@@ -1,4 +1,7 @@
 <?php
+
+use function PHPSTORM_META\elementType;
+
 session_start();
 include_once "../utils/routerConfig.php";
 include "../classes/database.php";
@@ -13,13 +16,20 @@ if (isset($_POST['category_name'])) {
     $category_name = $_POST['category_name'];
     try {
         $category = new Category(1, create_slug($category_name), $category_name);
-        Category::add($connection, $category);
-        echo '<script>alert("Thêm thể loại thành công!");
-            location.href = "' . BASE_URL . '/category";
-        </script>';
+        $is_exist = Category::getByCategory($connection, $category);
+        if (!is_object($is_exist)) {
+            Category::add($connection, $category);
+            echo '<script>alert("Thêm thể loại thành công!");
+                location.href = "' . BASE_URL . '/category";
+            </script>';
+        } else {
+            echo '<script>alert("Tên thể loại đã tồn tại vui lòng thử lại!");
+                location.href = "' . BASE_URL . '/category";
+            </script>';
+        }
     } catch (\Throwable $e) {
         echo '<script>alert("Đã xảy ra lỗi vui lòng thử lại!");
-            location.href = "' . BASE_URL . '/category";
+            
         </script>';
     }
 }
